refactor: 코드 스멜 제거, 코드 리뷰 반영
